% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_manipulation.R
\name{api_no_city}
\alias{api_no_city}
\title{api no city}
\usage{
api_no_city(
  db_name,
  dat,
  country,
  region = NULL,
  state = NULL,
  county = NULL,
  city = NULL,
  silent = FALSE
)
}
\arguments{
\item{db_name}{name of the SQLite database. If not exists, it will be created.}

\item{dat}{The tibble containing the data. It MUST contain a collumn called 'ID' with UNIQUE identification numbers
and at least names of the cities, country and date of registration of the Org.}

\item{country}{String with the name of the column with the country codes (2-letter)}

\item{region}{NOT RECOMENDED. String with the name of the column with the region names.}

\item{state}{Optional. String with the name of the column with the state names}

\item{county}{Optional. String with the name of the column with the county names}

\item{city}{As from version 2.3.1, this parameter is not necessary and it will be deprecated.}

\item{silent}{If TRUE, silences the messages from the function}
}
\value{
Creates a table (if not yet exists) called "orgs" in the given SQLite data base. The table contains the ID,
country code and the extra field(s) provided, together with the latitude and longitude of the region/state/county,
per organization. The column City also exists but is filled with empty values.
}
\description{
Performs the webs crapping ignoring city names
}
\details{
The function provides an alternative system to search for region or state coordinates, for example,
if the map has to be done per region instead of per city.
}
\author{
Manuel Teodoro
}
