% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_manipulation.R
\name{api_to_db}
\alias{api_to_db}
\title{api to database}
\usage{
api_to_db(
  mdb,
  dat,
  city = "City",
  country = "Country",
  region = NULL,
  state = NULL,
  county = NULL,
  db_backup_after = 10,
  silent = FALSE
)
}
\arguments{
\item{mdb}{Mapic database configuration object (See
\link{database_configuration} for a reference).}

\item{dat}{The data frame containing the data. It MUST contain a collumn called 'ID' with UNIQUE identification
numbers, and at least names of the cities and country.}

\item{city}{String with the name of the column with the city names}

\item{country}{String with the name of the column with the country codes (2-letter)}

\item{region}{String with the name of the column with the region names. NOT RECOMENDED.}

\item{state}{Optional. String with the name of the column with the state names}

\item{county}{Optional. String with the name of the column with the county names}

\item{db_backup_after}{Number of iteration after which the data is sent to the database.}

\item{silent}{If TRUE, silences the messages from the function}
}
\description{
Based on a list of countries, retrieve coordinates from OSM api and
send the results to a database or database alternative.
}
\details{
Given a list of cities and country(ies) in a data frame (as tibble),
the function iterates over the values to find the coordinates, using
the function \code{coords_from_city} and send the results to a database

The data must be provided as a data frame, the unique identification
numbers must be provided in a column called "ID". Other values that MUST be
provided are name of the cities and country in a 2-letter code.

The function uses \code{coords_from_city} for retreiving the data from the API,
therefore the values \code{state} and \code{county} are optional.
The results of the query are added to the database or database alternative.
When more than one result is found, only the first entry is added.
}
\author{
Manuel Teodoro
}
