% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_map.R
\name{base_map}
\alias{base_map}
\title{Base map}
\usage{
base_map(
  country,
  x_limits = NULL,
  y_limits = NULL,
  show_coords = FALSE,
  return_mapic_obj = TRUE,
  map_colors = default_map_colors
)
}
\arguments{
\item{country}{A string. Name of the country, as defined in the package \code{maps}.}

\item{x_limits}{A vector of size 2 containing the horizontal limits of the map.}

\item{y_limits}{A vector of size 2 containing the vertical limits of the map.}

\item{show_coords}{If \code{TRUE}, the coordinates are shown in the map. It is particularly
useful to set the x and y limits.}

\item{return_mapic_obj}{If \code{TRUE} (default), it returns an object of class \code{mapicHolder}.
Otherwise it returns a \code{ggplot} object.}

\item{map_colors}{Object of class \code{map_colors} containing the values for the colors
to use for the creation of the map. If not provided, it will use the \code{default_map_colors}.}
}
\value{
The function returns the map, either as a \code{mapicHolder} object or as a \code{ggplot}
(depending on \code{return_mapic_obj} flag).
If the object is of class \code{ggplot}, it can be used with other \code{ggplot} family of functions.
If the object is of class \code{mapicHolder} it contains embeded information to support
further steps in the mapping process.
The \code{ggplot} object is included in the \code{$mapic} element of the object.
}
\description{
Creates the base map of the country
}
\details{
Creates the base map of a country.
}
\examples{
base_map("Brazil")

base_map("Mexico",
         x_limits = c(-118, -86),
         y_limits = c(14, 34),
         return_mapic_obj = FALSE,
         show_coords = T) +
  scale_x_continuous(n.breaks = 20) +
  ggtitle("A map of Mexico")
}
