% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_connections.R
\name{combine_csv_sql}
\alias{combine_csv_sql}
\title{combine csv to sql table}
\usage{
combine_csv_sql(db_file, csv_file, city = "City")
}
\arguments{
\item{db_file}{Path to the SQLite file containing the table "orgs" generated by the function
\code{webscrap_to_sqlite}, or with the same parameters}

\item{csv_file}{Path to the csv file, or data.frame containing the source information. It MUST contain a
column with Unique Identification Numbers called "ID".}

\item{city}{String with the name of the column with the city names}
}
\value{
The function returns a tibble containing the joined information
}
\description{
Joins the data from the SQLite and CSV tables
}
\details{
Creates an inner join between the original data (either as csv or data frame tibble) and the data
contained in the database table. The joint is performed by ID and City fields. The fields Region, State and
County are taken from the database.
}
\author{
Manuel Teodoro
}
