% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_from_city.R
\name{coords_from_city}
\alias{coords_from_city}
\title{Coordinates from city}
\usage{
coords_from_city(
  city = NULL,
  country_code,
  region = NULL,
  state = NULL,
  county = NULL,
  choose_when_multiple = FALSE,
  silent = FALSE
)
}
\arguments{
\item{city}{Name of the city}

\item{country_code}{the 2-letter code of the country}

\item{region}{Optional, region name. This option is NOT RECOMENDED,
it's functioning is not wel docummented.}

\item{state}{Optional, state name.}

\item{county}{Optional, county name.}

\item{choose_when_multiple}{Not implemented yet!.}

\item{silent}{If TRUE, silences the messages from the function.}
}
\value{
Returns a 1-row data frame, containing latitude, longitude and osm name.
}
\description{
Get coordinates of a city per country.
}
\details{
Get the coordinates of a city, given the Country 2-letter code.
The search can be supported by region, state and/or county (optional).
The function uses open street maps nominatim api.

All the variables used by the function must be strings
}
