% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_configuration.R
\name{database_configuration}
\alias{database_configuration}
\title{database configuration}
\usage{
database_configuration(db_type, table, database)
}
\arguments{
\item{db_type}{String describing the type of database to use. Options are
"data.frame", "csv" and "SQLite".}

\item{table}{A string with the name of thetable. For data frame, the name of the object; for csv the path to the file; for SQL the table name.}

\item{database}{A string with the name of the database. Not necessary for csv and data frame.}
}
\value{
Returns an S3 object of different class (mapic db configuration object),
depending on the type of database selected.
}
\description{
Creates the corresponding mapic db configuration object containing the database configuration details.
}
\details{
The function returns and object of the corresponding class, depending of the database specified:
\itemize{
\item "data.frame" returns object of class "mdb_df". Here,
all the data obtained from the API
connection is stored internally in R memory. When the object \code{mdb_df} is
passed to \code{\link{db_load}}, it creates a data frame in the global environment
with the name specified in \code{table}. Therefore, it is important that an
object with the same name does not exist yet, or it will create a conflict.
\item "csv" returns object of class "mdb_csv". It uses a text-based csv file
to store the data obtained from the API connection. When the object
\code{mdb_csv} is passed to \code{\link{db_load}}, it will create a csv file in the
path specified in \code{table}. It is recommended that such file is not
created manually or it can generate conflicts.
\item "SQLite" returns object of class "mdb_SQLite". It uses SQLite as a
database. When the object \code{mdb_SQLite} is passed to \code{\link{db_load}},
it can generate both, the file and the table automatically, or if the file
already exists, it searches for the table. It is recommended that the table is
generated by the function and not manually.
It requires to have \href{https://www.sqlite.org/}{SQLite} installed in the system calling the function.
}
}
