% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_append.R
\name{db_append}
\alias{db_append}
\alias{db_append.default}
\alias{db_append.data.frame}
\alias{db_append.mdb_csv}
\alias{db_append.mdb_SQLite}
\title{Appends rows to the database table}
\usage{
db_append(x, ...)

\method{db_append}{default}(x, ...)

\method{db_append}{data.frame}(x, ...)

\method{db_append}{mdb_csv}(x, ...)

\method{db_append}{mdb_SQLite}(mdb, df)
}
\arguments{
\item{db_file}{Path to the SQLite file containing the table 'orgs'}
}
\value{
An object of class tibble (which inherits data.frame) containing the values stored
in the database with name \code{db_file}
}
\description{
Appends the rows of a data frame into the database table
}
\section{Methods (by class)}{
\itemize{
\item \code{db_append(default)}: Default

\item \code{db_append(data.frame)}: data.frame

\item \code{db_append(mdb_csv)}: mdb_csv

\item \code{db_append(mdb_SQLite)}: mdb_SQLite

}}
\author{
Manuel Teodoro
}
