% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_join_original_data.R
\name{db_join_original_data}
\alias{db_join_original_data}
\title{join the database to the original data}
\usage{
db_join_original_data(mdb, original_data, city = "City")
}
\arguments{
\item{mdb}{Mapic database configuration object (See
\link{database_configuration} for a reference).}

\item{city}{String with the name of the column with the city names.}

\item{df}{data frame with the original data. It MUST contain a
column with Unique Identification Numbers called "ID".}
}
\value{
The function returns a \code{data.frame} containing the joined data.
}
\description{
Joins the data from the database and the original data.
}
\details{
Creates an inner join between the original data and the data
contained in the database table. The joint is performed by "ID" and "City"
fields. The fields "Region", "State" and
"County" are taken from the database.
}
