% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_remove_empty.R
\name{db_remove_empty}
\alias{db_remove_empty}
\alias{db_remove_empty.default}
\alias{db_remove_empty.mdb_df}
\alias{db_remove_empty.mdb_csv}
\alias{db_remove_empty.mdb_SQLite}
\title{Removes empty entries from database table}
\usage{
db_remove_empty(mdb, ...)

\method{db_remove_empty}{default}(mdb)

\method{db_remove_empty}{mdb_df}(mdb)

\method{db_remove_empty}{mdb_csv}(mdb)

\method{db_remove_empty}{mdb_SQLite}(mdb)
}
\arguments{
\item{mdb}{Mapic database configuration object (See
\link{database_configuration} for a reference).}
}
\description{
Removes from the database table the entries where the \code{lat} and \code{lot} are empty.
}
\details{
When the function link{api_to_db} does not find an entry in the query,
it adds it to the database with the fields \code{lat} and \code{lon} empty. The
present function removes those entries from the database completely.
}
\section{Methods (by class)}{
\itemize{
\item \code{db_remove_empty(default)}: Default

\item \code{db_remove_empty(mdb_df)}: mdb_df

\item \code{db_remove_empty(mdb_csv)}: mdb_csv

\item \code{db_remove_empty(mdb_SQLite)}: mdb_SQLite

}}
