% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aesthetics.R
\name{define_legends_specs}
\alias{define_legends_specs}
\title{Define map legends specifications}
\usage{
define_legends_specs(legend_year, legend_orgs)
}
\arguments{
\item{legend_year, legend_orgs}{Exact text to be shown in the maps.}
}
\value{
An S3 object of class \code{map_legends} that inherits \code{map_aesthetics}.
}
\description{
Creates an object of class \code{map_legends} that contains the specifications used for the legends.
}
\examples{
my_legends_specs <- define_legends_specs("Year", "Organizations")
}
\author{
Manuel Teodoro
}
