% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aesthetics.R
\name{define_map_aesthetics}
\alias{define_map_aesthetics}
\title{Define map aesthetics}
\usage{
define_map_aesthetics(aes_name, colors, sizes, legends)
}
\arguments{
\item{aes_name}{A string with a name for the object.}

\item{colors}{Object of class \code{map_colors} defining the colors for the maps.}

\item{sizes}{Object of class \code{map_sizes} defining the sizes for the maps.}

\item{legends}{Object of class \code{map_legends} defining the aesthetics for the legends for the maps.}
}
\value{
An S3 object of class \code{map_master_aes}.
}
\description{
Creates an object of class \code{map_master_aes} that contains the aesthetics details for colors,
sizes and legends text.
}
\details{
The object inherits both classes, \code{map_aesthetics} and \code{map_master_aes}.
It works as an object to collect all the aesthetics defined for color, sizes and legends.
}
\examples{
my_map_colors <- define_map_colors(dots_orgs = "#574166",
                                   target_country = "#8caeb4",
                                   empty_countries = "#f3f3f3",
                                   border_countries = "#9c9c9c",
                                   oceans = "#4e91d2",
                                   text_cities = "#a0a0a0",
                                   text_legend = "#283151",
                                   text_copyright = "#f3f3f3")
my_sizes_in_map <- define_sizes_in_map(basic_dot = 2.5, text_city = 4)
my_legends_specs <- define_legends_specs("Year", "Organizations")

my_aesthetics <- define_map_aesthetics("My maps aesthetics",
                                        my_map_colors,
                                        my_sizes_in_map,
                                        my_legends_specs)
}
\author{
Manuel Teodoro
}
