% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_colors.R
\name{define_map_colors}
\alias{define_map_colors}
\title{Define map colors}
\usage{
define_map_colors(
  dots_orgs,
  target_country,
  empty_countries,
  border_countries,
  oceans,
  text_cities,
  text_legend,
  background_legend,
  text_copyright
)
}
\arguments{
\item{dots_orgs, target_country, empty_countries, border_countries, oceans, text_cities, text_legend, text_copyright}{String as character defining the colors of the different aspects of the maps, in hexadecimal notation.}
}
\value{
An S3 object of class \code{map_colors} that inherits \code{map_aesthetics}.
}
\description{
Creates an object of class \code{map_colors} that contains colors used for the maps.
}
\examples{
my_map_colors <- define_map_colors(dots_orgs = "#574166",
                                   target_country = "#8caeb4",
                                   empty_countries = "#f3f3f3",
                                   border_countries = "#9c9c9c",
                                   oceans = "#4e91d2",
                                   text_cities = "#a0a0a0",
                                   text_legend = "#283151",
                                   text_copyright = "#f3f3f3")
}
\author{
Manuel Teodoro
}
