% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aesthetics.R
\name{define_sizes_in_map}
\alias{define_sizes_in_map}
\title{Define map sizes}
\usage{
define_sizes_in_map(basic_dot, text_city)
}
\arguments{
\item{basic_dot, text_city}{Numeric values defining the sizes of the different aspects to be used in the maps.}
}
\value{
An S3 object of class \code{map_sizes} that inherits \code{map_aesthetics}.
}
\description{
Creates an object of class \code{map_sizes} that contains the sizes used for the maps.
}
\examples{
my_sizes_in_map <- define_sizes_in_map(basic_dot = 2.5, text_city = 4)
}
\author{
Manuel Teodoro
}
