% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_connections.R
\name{import_db_as_df}
\alias{import_db_as_df}
\title{Import database as data frame}
\usage{
import_db_as_df(db_file)
}
\arguments{
\item{db_file}{Path to the SQLite file containing the table 'orgs'}
}
\value{
An object of class tibble (which inherits data.frame) containing the values stored
in the database with name \code{db_file}
}
\description{
Imports the database as a tibble data frame
}
\author{
Manuel Teodoro
}
