% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aesthetics.R
\name{is.map_aesthetics}
\alias{is.map_aesthetics}
\title{Test if the object is of class \code{map_aesthetics}}
\usage{
is.map_aesthetics(x)
}
\arguments{
\item{x}{An object}
}
\value{
'TRUE' if the object inhertis from the {map_aesthetics} class
}
\description{
This function returns ‘TRUE’ for \code{map_aesthetics} objects and ‘FALSE’ for all other objects.
}
\author{
Manuel Teodoro
}
