% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aesthetics.R
\name{is.map_colors}
\alias{is.map_colors}
\title{Test if the object is of class \code{map_colors}}
\usage{
is.map_colors(x)
}
\arguments{
\item{x}{An object}
}
\value{
'TRUE' if the object inhertis from the \code{map_colors} class
}
\description{
This function returns ‘TRUE’ for \code{map_colors} objects and ‘FALSE’ for all other objects.
}
\author{
Manuel Teodoro
}
