% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aesthetics.R
\name{is.map_sizes}
\alias{is.map_sizes}
\title{Test if the object is of class \code{map_sizes}}
\usage{
is.map_sizes(x)
}
\arguments{
\item{x}{An object}
}
\value{
'TRUE' if the object inhertis from the \code{map_sizes} class
}
\description{
This function returns ‘TRUE’ for \code{map_sizes} objects and ‘FALSE’ for all other objects.
}
\author{
Manuel Teodoro
}
