% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_country_prev.R
\name{map_country_prev}
\alias{map_country_prev}
\title{Map country preview}
\usage{
map_country_prev(
  country,
  x_limits = NULL,
  y_limits = NULL,
  show_coords = FALSE,
  map_aes = default_aesthetics
)
}
\arguments{
\item{country}{A string. Name of the country, as defined in the package \code{maps}.}

\item{x_limits}{A vector of size 2 containing the horizontal limits of the map.}

\item{y_limits}{A vector of size 2 containing the vertical limits of the map.}

\item{show_coords}{If \code{TRUE}, the coordinates are shown in the map. It is particularly
useful to set the x and y limits.}

\item{map_aes}{Object of class \code{map_master_aes} containing the aesthetics required for
the creation of the map.}
}
\value{
The map.
}
\description{
Creates the base map of the country
}
\details{
Plots the map.
}
\examples{
map_country_prev("Brazil")
}
\author{
Manuel Teodoro
}
