% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapic_totals.R
\name{mapic_totals_internal}
\alias{mapic_totals_internal}
\title{Mapic totals}
\usage{
mapic_totals_internal(x, ...)
}
\arguments{
\item{.mapic_holder}{Object of class \code{mapicHolder} as returned from the function
\code{base_map} and \code{mapic_city_dots} (optional).}

\item{totals}{The value to be included, as \code{numeric}.
Not necessary if an object of class \code{mapicHolder} is passed.}

\item{x_limits}{A vector of size 2 containing the horizontal limits of the map.
Not necessary for external or if an object of class \code{mapicHolder} is passed.}

\item{y_limits}{A vector of size 2 containing the vertical limits of the map.
Not necessary for external or if an object of class \code{mapicHolder} is passed.}

\item{totals_label}{A string to be used for the label "Totals".}

\item{map_colors}{An object of class \code{map_colors} containing the details of the colors for the maps.
Not necessary if an object of class \code{mapicHolder} is passed.}
}
\value{
If the map is creating using an object of class \code{mapicHolder}, it returns the same object with added
information and map elements. Otherwise, it returns a \code{ggplot} object.
}
\description{
Generates the label of the total elements included in the map,
either as internal or external label, depending which function is being called.
}
\details{
It generates the labels to show which years is being mapped.
The internal version currently plots only on the lower-left corner of the map.
The external version generates a sepparated plot that has to be called sepparately.
}
\author{
Manuel Teodoro
}
