% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mexico}
\alias{mexico}
\title{Fake data simulating organizations from Mexico}
\format{
A data frame with 10 rows and 9 columns:
\describe{
\item{id}{Unique ID}
\item{name}{Name of the organization}
\item{type}{The type of organization}
\item{registration_year}{Year of its official registration}
\item{end_year}{Year of clossing or ending, if any}
\item{country}{Two letter code of the country}
\item{region}{The state of the country}
\item{city}{The city where it is registered}
\item{source}{The spurce of the data}
}
}
\usage{
mexico
}
\description{
A fake dataset simulating organizations of Mexico to test the functions.
It results in simple and fast queries and maps.
}
\keyword{datasets}
