% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_connections.R
\name{remove_na_from_db}
\alias{remove_na_from_db}
\title{remove na from database}
\usage{
remove_na_from_db(db_file)
}
\arguments{
\item{db_file}{Path to the SQLite file containing the table "orgs"}
}
\description{
Removes values not found from the database
}
\details{
Removes all the values that were not found in the OSM query and thus contain empty values of
\code{lat} and \code{lon}. It is IMPORANT to consider that the function modifes the database directly from
an SQL query.
}
\author{
Manuel Teodoro
}
