% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_manipulation.R
\name{webscrap_to_db}
\alias{webscrap_to_db}
\title{webscrap to database}
\usage{
webscrap_to_db(
  db_name,
  dat,
  city = "City",
  country = "Country",
  region = NULL,
  state = NULL,
  county = NULL,
  db_backup_after = 10
)
}
\arguments{
\item{db_name}{name of the SQLite database. If not exists, it will be created}

\item{dat}{The tibble containing the data. It MUST contain a collumn called 'ID' with UNIQUE identification
numbers, and at least names of the cities, country and date of registration of the Org.}

\item{city}{String with the name of the column with the city names}

\item{country}{String with the name of the column with the country codes (2-letter)}

\item{region}{String with the name of the column with the region names. NOT RECOMENDED.}

\item{state}{Optional. String with the name of the column with the state names}

\item{county}{Optional. String with the name of the column with the county names}
}
\value{
Creates a table (if not yet exists) called "orgs" in the given SQLite data base. The table contains
the ID, city name and country code obtained for the data inside the tibble \code{dat}, together with the
latitude and longitude of the city, per organization.
}
\description{
Retrieve coordinates from a list of countries and send it to a sql DB
}
\details{
Given a list of cities and country(ies) in a data frame (as tibble),
the function iterates over the values to find the coordinates, using
the function \code{coords_from_city} and send the results to a SQLite DB

The data must be provided as a tibble, the unique identification numbers must be provided in a
column called "ID". Other values that MUST be provided in the tibble are name of the cities and country in
a 2-letter code.

The function uses \code{coords_from_city} for the web scrapping, therefore the values \code{state} and
\code{county} are optional.
The results of the query will be added to a table called "orgs" into the SQLite database ONLY when a single
result is found. If more than one result is found, the first entry will be added, if none are found, the
details will be printed on screen and the entry will not be added to the SQLite table.

NOTE: As of version 2.3.1, the function passes to the DB empty strings \code{""} from the empty values of
state, region and county. This behaviour doesn't seem to conflict with the creation of the maps.
If necessary, it is easy to convert all such values to \code{NA} in R.
If bugs are found regarding this behaviour conflicting with the database, please report it immediately.
}
\author{
Manuel Teodoro
}
