% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_aesthetics.R
\name{with_default_aesthetics}
\alias{with_default_aesthetics}
\title{With default map aesthetics}
\usage{
with_default_aesthetics(colors = NA, sizes = NA, legends = NA)
}
\arguments{
\item{colors}{A list containing the parameters of the colors to be changed as passed in the function \code{\link{define_map_colors}}.}

\item{sizes}{A list containing the parameters of the sizes to be changed as passed in the function \code{\link{define_map_sizes}}.}

\item{legends}{A list containing the textfor the legends to be changed, as passed in the function \code{\link{define_map_legends}}.}
}
\description{
Modifies the default aesthetics.
}
\details{
The function modifies the global objects containing the default aesthetics.
}
\examples{
my_aesthetics <- with_default_aesthetics(list(dots_orgs = "#000000"), list(text_city = 12))
}
\seealso{
\code{\link{define_map_colors}} for details about the colors, \code{\link{define_map_sizes}} for details about the sizes and \code{\link{define_map_legends}} for details about the legends.
}
\author{
Manuel Teodoro
}
