% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_colors.R
\name{with_default_colors}
\alias{with_default_colors}
\title{With default map colors}
\usage{
with_default_colors(list_colors = NA)
}
\arguments{
\item{list_colors}{A list containing the parameters of the colors to be changed as passed in the function \code{\link{define_map_colors}}.}
}
\description{
Modifies the default colors
}
\details{
The function modifies the global objects containing the default colors.
}
\examples{
my_colors <- with_default_colors(list(dots_orgs = "#000000")
}
\seealso{
\code{\link{define_map_colors}} for details about the colors.
}
