% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_coords_manually.R
\name{add_coords_manually}
\alias{add_coords_manually}
\title{Add coordinates manually}
\usage{
add_coords_manually(complementary_data, mdb)
}
\arguments{
\item{complementary_data}{file containing the missing values. The dataset
providing the values must contain the exact same fields as the ones in the
database. Fields can be empty except for "id", "lat" and "lon".}

\item{mdb}{Mapic database configuration object (See
\link{database_configuration} for a reference).}
}
\value{
It appends the provided values to the database.
}
\description{
Add missing values manually, from as csv file or data frame to
the database or database alternative.
}
\details{
Note that if one missing city appears several times, you can provide
the lat and long via this function only once, making sure that all the fields
for this city match, and then run again
\code{link{api_to_sqlite}} which will find it in the database and set it for
all entries with the same city.
}
