% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_manipulation.R
\name{api_no_city}
\alias{api_no_city}
\title{api no city}
\usage{
api_no_city(
  mdb,
  dat,
  country,
  region = NULL,
  state = NULL,
  county = NULL,
  start_year = NULL,
  end_year = NULL,
  silent = FALSE
)
}
\arguments{
\item{mdb}{Mapic database configuration object (See
\link{database_configuration} for a reference).}

\item{dat}{The data frame containing the data. It MUST contain a collumn
named 'ID' with UNIQUE identification numbers, the country
and at least one of the following fields: state, county and/or region.}

\item{country}{String with the name of the column specifying the country
(2-letter code)}

\item{region}{NOT RECOMENDED. String with the name of the column with the region names.}

\item{state}{Optional. String with the name of the column with the state names}

\item{county}{Optional. String with the name of the column with the county names}

\item{start_year}{String with the name of the column with the year of start.}

\item{end_year}{String with the name of the column with the end.}

\item{silent}{If TRUE, silences the messages from the function}
}
\value{
The function fulfills the same task as \code{\link{api_to_db}}.
}
\description{
Retrieves data from the API ignoring city names.
}
\details{
The function provides an alternative system to search for region or
state coordinates, for example, if the map has to be done per region instead of
per city.
}
\seealso{
\link{api_to_db}.
}
\author{
Manuel Teodoro
}
