% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_append.R
\name{db_append}
\alias{db_append}
\alias{db_append.default}
\alias{db_append.mdb_df}
\alias{db_append.mdb_csv}
\alias{db_append.mdb_SQLite}
\title{Appends rows to the database table}
\usage{
db_append(mdb, ...)

\method{db_append}{default}(mdb, ...)

\method{db_append}{mdb_df}(mdb, df)

\method{db_append}{mdb_csv}(x, ...)

\method{db_append}{mdb_SQLite}(mdb, df)
}
\arguments{
\item{mdb}{Mapic database configuration object (See
\link{database_configuration} for a reference).}

\item{df}{data frame with the rows to be appended.}
}
\description{
Appends the rows of a data frame into the database table
}
\details{
The function appends all the rows in \code{df} to the database and
table specified in \code{mdb}.
}
\section{Methods (by class)}{
\itemize{
\item \code{db_append(default)}: Default

\item \code{db_append(mdb_df)}: mdb_df

\item \code{db_append(mdb_csv)}: mdb_csv

\item \code{db_append(mdb_SQLite)}: mdb_SQLite

}}
\seealso{
\link{database_configuration}
}
\author{
Manuel Teodoro
}
