% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_compare_data.R
\name{db_compare_data}
\alias{db_compare_data}
\title{Compares the database to the original data}
\usage{
db_compare_data(mdb, df)
}
\arguments{
\item{mdb}{Mapic database configuration object (See
\link{database_configuration} for a reference).}

\item{df}{data frame with the original data.}
}
\value{
An object of class \code{data.frame} containing the missing values
in the database (specified in \code{mdb}) as compared with the original \code{df}
}
\description{
Creates a data frame with the missing data.
}
\seealso{
\link{database_configuration}
}
