% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_load.R
\name{db_load}
\alias{db_load}
\alias{db_load.default}
\alias{db_load.mdb_df}
\alias{db_load.mdb_csv}
\alias{db_load.mdb_SQLite}
\alias{db_load.mdb_PostgreSQL}
\title{Loads the database table}
\usage{
db_load(mdb, ...)

\method{db_load}{default}(mdb)

\method{db_load}{mdb_df}(mdb)

\method{db_load}{mdb_csv}(mdb)

\method{db_load}{mdb_SQLite}(mdb)

\method{db_load}{mdb_PostgreSQL}(mdb)
}
\arguments{
\item{mdb}{Mapic database configuration object (See
\link{database_configuration} for a reference).}
}
\value{
An object of class \code{data.frame} containing the values stored
in the database specified in \code{mdb}.
}
\description{
Loads the database table to R as a data frame.
}
\details{
The main task of the function is to load the table specified in the
\code{mdb} object to R as a \code{data.frame}. The function also has the tasks of
creating the table when it does not exist yet. This is the equivalent to the
command \code{CREATE TABLE IF NOT EXISTS} in SQL. For \code{csv} it creates
the file and for \code{data.frame} it creates the data frame in the global
environment, both with the structure required by mapic to store the coordinates.
}
\section{Methods (by class)}{
\itemize{
\item \code{db_load(default)}: Default

\item \code{db_load(mdb_df)}: mdb_df

\item \code{db_load(mdb_csv)}: mdb_csv

\item \code{db_load(mdb_SQLite)}: mdb_SQLite

\item \code{db_load(mdb_PostgreSQL)}: mdb_PostgreSQL

}}
\seealso{
\link{database_configuration}
}
