% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapic_totals.R
\name{mapic_totals_internal}
\alias{mapic_totals_internal}
\alias{mapic_totals_internal.default}
\alias{mapic_totals_internal.mapicHolder}
\alias{mapic_totals_external}
\alias{mapic_totals_external.default}
\alias{mapic_totals_external.mapicHolder}
\title{Mapic totals generic}
\usage{
mapic_totals_internal(x, ...)

\method{mapic_totals_internal}{default}(
  totals,
  x_limits,
  y_limits,
  totals_label = "Totals",
  map_colors = default_map_colors
)

\method{mapic_totals_internal}{mapicHolder}(.mapic_holder, totals_label = "Totals")

mapic_totals_external(x, ...)

\method{mapic_totals_external}{default}(
  totals,
  totals_label = "Totals",
  map_colors = default_map_colors
)

\method{mapic_totals_external}{mapicHolder}(.mapic_holder, totals_label = "Totals")
}
\arguments{
\item{totals}{The value to be included, as \code{numeric}.
Not necessary if an object of class \code{mapicHolder} is passed.}

\item{x_limits}{A vector of size 2 containing the horizontal limits of the map.
Not necessary for external or if an object of class \code{mapicHolder} is passed.}

\item{y_limits}{A vector of size 2 containing the vertical limits of the map.
Not necessary for external or if an object of class \code{mapicHolder} is passed.}

\item{totals_label}{A string to be used for the label "Totals".}

\item{map_colors}{An object of class \code{map_colors} containing the details of the colors for the maps.
Not necessary if an object of class \code{mapicHolder} is passed.}

\item{.mapic_holder}{Object of class \code{mapicHolder} as returned from the function
\code{base_map} and \code{mapic_city_dots} (optional).}
}
\value{
If the map is created using an object of class \code{mapicHolder}, it returns the same object with added
information and map elements. Otherwise, it returns a \code{ggplot} object.
}
\description{
Generates the label of the total elements included in the map,
either as internal or external label, depending which function is being called.
}
\details{
It generates the labels to show the total values mapped.
The internal version currently plots only on the lower-left corner of the map.
The external version generates a sepparated plot that has to be called sepparately.
}
\section{Methods (by class)}{
\itemize{
\item \code{mapic_totals_internal(default)}: Default

\item \code{mapic_totals_internal(mapicHolder)}: mapicHolder

}}
\author{
Manuel Teodoro
}
