% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mexico}
\alias{mexico}
\title{Fake data simulating organizations from Mexico}
\format{
A data frame with 10 rows and 9 columns:
\describe{
\item{ID}{Unique ID}
\item{Name}{Name of the organization}
\item{Type}{The type of organization}
\item{Registration_year}{Year of its official registration}
\item{End_year}{Year of clossing or ending, if any}
\item{Country}{Two letter code of the country}
\item{Region}{The state of the country}
\item{City}{The city where it is registered}
\item{Source}{The spurce of the data}
}
}
\usage{
mexico
}
\description{
A fake dataset simulating organizations of Mexico to test the functions.
It results in simple and fast queries and maps.
}
\keyword{datasets}
